Moved things to folder, will make organization easier
