+++ Frontend/src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 // import logo from './logo.svg';
 import './App.css';
-import AccountHome from './components/AccountHome';
+import AccountHome from './components/accountPage/AccountHome';
 import Login from './components/Login';
 import WorkoutGenerator from './components/WorkoutGenerator';
 import WorkoutPage from './components/WorkoutPage';
