+++ Frontend/src/components/accountPage/AccountHome.jsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
-import Navigation from './Navigation';
+import Navigation from '../Navigation';
 import { Container, Row, Col, Button, Card } from 'react-bootstrap';
 import './AccountHome.css';
-import Account from '../models/Account';
-import { HomeRepository } from './../api/HomeRepository';
+import Account from '../../models/Account';
+import { HomeRepository } from '../../api/HomeRepository';
 
 class AccountHome extends Component {
